requirements.txt
requests>=2.25.0
openai>=1.0.0
google-generativeai>=0.3.0